chore: exclude mayush-mobile from repo and update gitignore
- Add mayush-mobile/ to .gitignore (mobile project managed separately)
- Add .codex-temp/, reports/, docs/superpowers/, OS files to .gitignore
- Un-track .claude/settings.local.json and .phpunit.cache/test-results
- Include pending Laravel backend changes (auth, migrations, inspiration, CORS)

+++ database/migrations/2026_05_07_000001_add_notification_type_id_to_notifications_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
         }
 
         Schema::table('notifications', function (Blueprint $table) {
-            $table->unsignedBigInteger('notification_type_id')->nullable()->after('id');
+            $table->unsignedInteger('notification_type_id')->nullable()->after('id');
         });
     }
